build: harden warnings and add unit test suite
- Enable /W4 /permissive- /w14062 on AmethystTool, isolate third-party via /external:W0
- Add CMakePresets (ninja-multi, vs2022, build/test presets), .clang-tidy, .clang-format
- Integrate GoogleTest v1.17.0 via FetchContent, gtest_force_shared_crt=OFF
- Add 33 tests covering parsers (Numbers, Encoding, Hash, Fnv1a)
- Defer 13 warnings (C4100/C4505); document /analyze as future work

+++ src/Hook/Hooks_Package.cpp
@@ -160,7 +160,7 @@ namespace Hooks_Package {
         LOG_PACKAGE_DEBUG("NotifyLicenseChanged: processing {} additions", additions.size());
         if (!additions.empty()) {
             uint32_t oldSize = pPkg->AppIdVec.m_Size;
-            if (CUtlMemoryGrowWrap(&pPkg->AppIdVec, additions.size())) {
+            if (CUtlMemoryGrowWrap(&pPkg->AppIdVec, static_cast<int>(additions.size()))) {
                 // An applied addition invalidates any UI removal that has not
                 // reached the UI thread yet.
                 for (AppId_t id : additions)
